Adds Manager canRaiseSalary Test Passing

+++ src/main/java/staff/management/Manager.java
@@ -6,7 +6,7 @@ public class Manager extends Employee {
 
     private String deptName;
 
-    public Manager(String name, int niNumber, int salary, String deptName) {
+    public Manager(String name, int niNumber, double salary, String deptName) {
         super(name, niNumber, salary);
         this.deptName = deptName;
     }
